@@ -1,14 +1,14 @@
 package com.lossydragon.arduinorgb.fragment
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.lossydragon.arduinorgb.R
 
 
-class FragmentHue : Fragment() {
+class FragmentHue : androidx.fragment.app.Fragment() {
 
     companion object {
         fun newInstance(): FragmentHue {
